- ad loaded event for interstitial ad - making sure that we have control on ad loading
