+++ MediaNetAdSDK/src/main/java/com/medianet/android/adsdk/BannerAd.kt
@@ -1,8 +1,14 @@
 package com.medianet.android.adsdk
 
+import com.google.android.gms.ads.AdListener
 import com.google.android.gms.ads.AdSize
-import org.prebid.mobile.AdUnit
+import com.google.android.gms.ads.LoadAdError
+import com.google.android.gms.ads.admanager.AdManagerAdRequest
+import com.google.android.gms.ads.admanager.AdManagerAdView
+import com.medianet.android.adsdk.utils.Util
 import org.prebid.mobile.BannerAdUnit
+import org.prebid.mobile.addendum.AdViewUtils
+import org.prebid.mobile.addendum.PbFindSizeError
 
 
 class BannerAd(adUnitId: String, val adSize: AdSize = AdSize.BANNER): Ad(BannerAdUnit("imp-prebid-banner-300-250", adSize.width, adSize.height)) {
@@ -20,5 +26,58 @@ class BannerAd(adUnitId: String, val adSize: AdSize = AdSize.BANNER): Ad(BannerA
 
     fun addAdditionalSize(width: Int, height: Int) = apply {
         bannerAdUnit.addAdditionalSize(width, height)
+    }
+
+    fun fetchDemandAndLoad(
+        adView: AdManagerAdView,
+        adRequest: AdManagerAdRequest,
+        listener: GamEventListener
+    ) {
+        adView.adListener = object : AdListener() {
+            override fun onAdLoaded() {
+                // Update ad view
+                AdViewUtils.findPrebidCreativeSize(adView, object : AdViewUtils.PbFindSizeListener {
+                    override fun success(width: Int, height: Int) {
+                        adView.setAdSizes(AdSize(width, height))
+                    }
+
+                    override fun failure(error: PbFindSizeError) {
+                    }
+                })
+                sendAdLoadedEvent()
+                listener.onAdLoaded()
+            }
+
+            override fun onAdClicked() {
+                listener.onAdClicked()
+            }
+
+            override fun onAdClosed() {
+                listener.onAdClosed()
+            }
+
+            override fun onAdFailedToLoad(p0: LoadAdError) {
+                listener.onAdFailedToLoad(Util.mapGamLoadAdErrorToError(p0))
+            }
+
+            override fun onAdOpened() {
+                listener.onAdOpened()
+            }
+
+            override fun onAdImpression() {
+                listener.onAdImpression()
+            }
+        }
+
+        fetchDemand(adRequest, object : OnBidCompletionListener{
+            override fun onSuccess(keywordMap: Map<String, String>?) {
+                listener.onSuccess()
+                adView.loadAd(adRequest)
+            }
+
+            override fun onError(error: Error) {
+                listener.onAdFailedToLoad(error)
+            }
+        })
     }
 }
