+++ MediaNetAdSDK/src/main/java/com/medianet/android/adsdk/GamEventListener.kt
@@ -1,7 +1,10 @@
 package com.medianet.android.adsdk
 
+import com.google.android.gms.ads.admanager.AdManagerInterstitialAd
+
 interface GamEventListener: OnBidCompletionListener {
     fun onAdLoaded(){}
+    fun onInterstitialAdLoaded(ad: AdManagerInterstitialAd){}
     fun onAdClicked(){}
     fun onAdClosed(){}
     fun onAdFailedToLoad(error: Error){}
